+++ api/app/Http/Controllers/Web/RentAreaController.php
@@ -6,11 +6,15 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\RentAreaStoreRequest;
 use App\Models\RentArea;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
+use Symfony\Component\HttpFoundation\Response;
 
 class RentAreaController extends Controller
 {
     public function index()
     {
+        abort_if(Gate::denies('manage_rent_area'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         $rentAreas = RentArea::all();
 
         return view('rentarea.index', compact('rentAreas'));
@@ -18,6 +22,8 @@ class RentAreaController extends Controller
 
     public function create()
     {
+        abort_if(Gate::denies('manage_rent_area'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         return view('rentarea.create');
     }
 
@@ -30,6 +36,8 @@ class RentAreaController extends Controller
 
     public function edit(RentArea $rentArea)
     {
+        abort_if(Gate::denies('manage_rent_area'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         return view('rentarea.edit', compact('rentArea'));
     }
 
@@ -42,6 +50,8 @@ class RentAreaController extends Controller
 
     public function destroy(RentArea $rentArea)
     {
+        abort_if(Gate::denies('manage_rent_area'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         $rentArea->delete();
 
         return redirect()->route('rentarea.index');
